+++ makefiles/prog/ppc/mem.h
@@ -5,7 +5,7 @@
 
 # include "./types.h"
 
-void CPROC *AllocateEx( int nSize DBG_PASS );
+void CPROC *AllocateEx( size_t nSize DBG_PASS );
 # define Allocate(s) AllocateEx(s DBG_SRC)
 void CPROC ReleaseExx( void ** DBG_PASS );
 
@@ -24,8 +24,8 @@ void DumpMemory( void );
 
 uint32_t CPROC LockedExchange( uint32_t *p, uint32_t val );
 
-void CPROC MemSet( POINTER p, uint32_t v, uint32_t n);
-void CPROC MemCpy( POINTER p, POINTER p2, uint32_t n);
+void CPROC MemSet( POINTER p, uint32_t v, size_t n);
+void CPROC MemCpy( POINTER p, POINTER p2, size_t n);
 
 
 void DisableMemoryValidate( int bDisable );
